refactor: rework app directory structure
