+++ app/imports/components
@@ -1,0 +1,1 @@
+../../src/raccoon_spotter/models/components/
