+++ app/imports/utils
@@ -1,0 +1,1 @@
+../../src/raccoon_spotter/utils
