--- app/models/README.md
@@ -1,1 +1,0 @@
-put trained models from TensorFlow/Keras here.
