+++ app/requirements.txt
@@ -4,4 +4,6 @@ tensorflow~=2.16.1
 numpy~=1.23.5
 Pillow~=9.5.0
 gevent~=22.10.2
-opencv-python ~= 4.8.0
+opencv-python~=4.8.0
+seaborn~=0.12.1
+mplcyberpunk~=0.7.1
