+++ src/raccoon_spotter/utils/__init__.py
@@ -1,1 +1,1 @@
-from .configs import configs, loader  # NOQA
+# from .configs import configs, loader  # NOQA
