+++ src/raccoon_spotter/utils/data_conversion.py
@@ -1,0 +1,26 @@
+import base64
+import re
+from io import BytesIO
+
+from PIL import Image
+
+
+def base64_to_pil(img_base64):
+    """
+    Convert base64 image data to PIL image
+    """
+    image_data = re.sub("^data:image/.+;base64,", "", img_base64)
+    pil_image = Image.open(BytesIO(base64.b64decode(image_data)))
+    return pil_image
+
+
+def np_to_base64(img_np):
+    """
+    Convert numpy image (RGB) to base64 string
+    """
+    img = Image.fromarray(img_np.astype("uint8"), "RGB")
+    buffered = BytesIO()
+    img.save(buffered, format="PNG")
+    return "data:image/png;base64," + base64.b64encode(buffered.getvalue()).decode(
+        "ascii"
+    )
